Add local stories linking feature
Stories in the same Project can now be referenced on another story's
description. A new parser was added on the StoryView to intercept and
turn '#<story_id>' into links which would later turn into StoryLinkViews
providing the current state of the story.

Add popover to story links

Refactor StoryLink code into react components

StoryLink react component was added to StoryView as well as a new StoryDescription
component

Fix react components filenames and user two-step parser on description

In order to be able to have onClick events on StoryLink, the parser had to
be split into two different steps.

- The markdown is parsed and a regexp is used to find and validate
local story ids on the description

- On StoryLink those elements are tranformed into a StoryLink component.