feature(story_state): set story state as unstarted
Set an estimable story as unstarted if it gets estimated and set it
back to unscheduled if it gets unestimated.

An estimable story can't have a state of unstarted while it's not
estimated and can't have a state of unscheduled while it's estimated.

So far only the feature stories are estimable.