Unstarded story state (#578)
* feature(story_state): set story state as unstarted

Set an estimable story as unstarted if it gets estimated and set it
back to unscheduled if it gets unestimated.

An estimable story can't have a state of unstarted while it's not
estimated and can't have a state of unscheduled while it's estimated.

So far only the feature stories are estimable.

* chore(stories_spec): set story estimate to a value

An unestimated story can't change its state to unstarted which was
causing the test to fail. Setting an estimate to the story fixes the
problem.

* test(stories_spec): fix unable to find element

With the new change that the estimated story is moved to the current
sprint column automatically, the story button `start` is not available
in the chilly bin column anymore.

* refactor(story_form): remove unused form

This story form was used to render the older collapsed stories view and
was forgotten to be removed. Nowadays the collapsed stories is a React
component and this form is not needed anymore.

* chore(gemfile): update central support gem

Point the central support gem to the last commit